Добавил обновление актуалььного рейтинга TTW через сутки + изменил правила
