+++ build.gradle
@@ -26,6 +26,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-test'
     implementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation 'org.projectlombok:lombok:1.18.22'
     compileOnly 'org.projectlombok:lombok'
 	runtimeOnly 'mysql:mysql-connector-java'
 	annotationProcessor 'org.projectlombok:lombok'
@@ -45,6 +46,10 @@ dependencies {
 	implementation 'org.liquibase:liquibase-core:4.7.1'
 	implementation 'org.modelmapper:modelmapper:3.1.1'
 	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-mail', version: '2.7.1'
+	// https://mvnrepository.com/artifact/org.jsoup/jsoup
+	implementation group: 'org.jsoup', name: 'jsoup', version: '1.15.4'
+
+
 
 	//		<=========== Swagger ===========>
 	implementation 'org.springdoc:springdoc-openapi-common:1.6.14'
