+++ src/main/java/site/tt_nsk/controller/ErrorController.java
@@ -6,6 +6,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class ErrorController {
     @GetMapping("/access-denied")
     public String accessDenied() {
-        return "access-denied";
+        return "error/access-denied";
     }
 }
