+++ src/main/java/site/tt_nsk/controller/PlayController.java
@@ -40,15 +40,16 @@ import java.util.stream.Collectors;
 @RequestMapping("/tour")
 public class PlayController {
     private final PlayService playService;
+    private final LegUpService legUpService;
     private final TourController tourController;
     private final PlayerService playerService;
     private final AddressService addressService;
     private final TourImageService tourImageService;
     private final TourDao tourDao;
-    Map<String, Scoring> resultTour;
-    String scoring;
-    List<String> list;
-    String filename;
+    static Map<String, Scoring> resultTour;
+    static String scoring;
+    static List<String> list;
+    static String filename;
     private static final String path = "tours";
     @Value("${storage.location}")
     private String storagePath;
@@ -92,7 +93,7 @@ public class PlayController {
         if (id != null) {
             tour = tourDao.findById(id).get();
             allByRating = tourService.getListPlayersForFutureTour(tourService.findAllByTourId(id));
-            legUp = playService.getLegUp(playService.getLegUpBeforeStartingTour(playService.getCurrentRatingAllPlayers(allByRating), allByRating));
+            legUp = legUpService.getLegUp(legUpService.getLegUpBeforeStartingTour(playService.getCurrentRatingAllPlayers(allByRating), allByRating));
             list = playService.arrayWithoutNull(playService.getListResultTour(score));
             score.setEndTour((playService.getSizeArrayList(list)/allByRating.size() + 1) == allByRating.size());
             resultTour = playService.getResultTour(list, allByRating);
@@ -102,7 +103,7 @@ public class PlayController {
         } else {
             tour = new Tour();
             allByRating = playerService.findAllActiveSortedByRating();
-            legUp = playService.getLegUp(playService.getLegUpBeforeStartingTour(playService.getCurrentRatingAllPlayers()));
+            legUp = legUpService.getLegUp(legUpService.getLegUpBeforeStartingTour(playService.getCurrentRatingAllPlayers()));
             list = playService.arrayWithoutNull(playService.getListResultTour(score));
             score.setEndTour((playService.getSizeArrayList(list)/allByRating.size() + 1) == allByRating.size());
             resultTour = playService.getResultTour(list);
@@ -158,39 +159,7 @@ public class PlayController {
         }
     }
 
-    public HashMap<String, String> getLegUpBeforeStartingTour(List<Double> currentRating){
-        HashMap<String, String> legUpStrArr = new HashMap<>();
-        Double currentRatingElement = getAllActiveSortedByRating().get(0).getRating().doubleValue();
-        for (int i = 0; i < currentRating.size() - 1 ; i++) {
-            for (int j = 1; j < currentRating.size(); j++) {
-                if (currentRatingElement > currentRating.get(j) && currentRatingElement != 500) {
-                    legUpStrArr.put("fx" + (i + 1) + "y" + (j + 1), scoringLegUp(currentRatingElement, currentRating.get(j)));
-                }
-            }
-            currentRatingElement = currentRating.get(i + 1);
-        }
-        return legUpStrArr;
-    }
 
-    public String scoringLegUp (Double ratingPlayerHighRating, Double ratingPlayerLowRating){
-        double difference = ratingPlayerHighRating - ratingPlayerLowRating;
-        if (difference >= 0 && difference <= 25){
-            return "0/0";
-        } else if (difference > 25 && difference <= 50){
-            return "0/1";
-        } else if (difference > 50 && difference <= 75){
-            return "0/2";
-        } else if (difference > 75 && difference <= 100){
-            return "0/3";
-        } else if (difference > 100 && difference <= 125){
-            return "0/4";
-        } else if (difference > 125 && difference  <= 150){
-            return "0/5";
-        } else if (difference > 150 && difference <= 175) {
-            return "0/6";
-        }
-        return "0/7";
-    }
     public Tour savePlayedTour(Tour tour, MultipartFile multipartFile) {
         tour.setAmountPlayers(BigDecimal.valueOf(resultTour.size()));
         tour.setDate(new Date());
